move tests to pytorch
