+++ python/nano/test/pytorch/test_trainer.py
@@ -21,8 +21,8 @@ from unittest import TestCase
 import pytest
 import torch
 from pytorch_lightning import LightningModule
-from test._train_torch_lightning import create_data_loader, data_transform
-from test._train_torch_lightning import train_with_linear_top_layer
+from test.pytorch.utils._train_torch_lightning import create_data_loader, data_transform
+from test.pytorch.utils._train_torch_lightning import train_with_linear_top_layer
 from torch import nn
 
 from bigdl.nano.pytorch.trainer import Trainer
