+++ python/nano/test/pytorch/vision/test_imagefolder.py
@@ -19,7 +19,7 @@ import pytest
 import os
 from unittest import TestCase
 from bigdl.nano.pytorch.vision.models import vision
-from test._train_imagefolder import train_torch_lightning
+from test.pytorch.utils._train_imagefolder import train_torch_lightning
 
 
 batch_size = 1
@@ -28,7 +28,7 @@ root_dir1 = os.path.join(resources_root, "train_image_folder_png")
 root_dir2 = os.path.join(resources_root, "train_image_folder_jpg")
 
 
-class TestModelsVision(TestCase):
+class TestImageFolder(TestCase):
 
     def test_resnet18_quantitrain_image_folder_pngze(self):
         resnet18 = vision.resnet18(
